--- src/components/pages/adminpage/cars/modals/CarDeleteModal/CarDeleteModal.tsx
@@ -1,26 +1,0 @@
-import { Modal } from "antd";
-import { Car } from "../../../../../../types/Car";
-import CarStore from "../../../../../../store/CarStore";
-
-interface ICarDeleteModal {
-    setVisible: (value: boolean) => void;
-    car: Car;
-}
-
-function CarDeleteModal({ setVisible, car }: ICarDeleteModal) {
-    return (
-        <Modal title="Вы уверены что хотите удалить этот автомобиль?"
-            open={true}
-            okText="Да"
-            cancelText="Нет"
-            onOk={async () => {
-                await CarStore.deleteCar(car);
-                setVisible(false);
-            }}
-            onCancel={() => setVisible(false)}
-        >
-        </Modal>
-    );
-}
-
-export default CarDeleteModal;
