--- src/components/share/Layout/Layout.tsx
@@ -1,16 +1,0 @@
-import IMain from "../../../interfaces/IMain";
-import Header from "../../navpanel/Header/Header";
-import classes from "./Layout.module.css";
-
-function Layout({ children }: IMain) {
-    return (
-        <div>
-            <Header />
-            <div className={classes.content}>
-                {children}
-            </div>
-        </div>
-    );
-}
-
-export default Layout;
