+++ src/components/shared/Layout/Layout.tsx
@@ -1,0 +1,17 @@
+import IMain from "interfaces/IMain";
+import Header from "@shared/Header/Header";
+import classes from "./Layout.module.css";
+
+
+function Layout({ children }: IMain) {
+    return (
+        <div>
+            <Header />
+            <div className={classes.content}>
+                {children}
+            </div>
+        </div>
+    );
+}
+
+export default Layout;
