+++ src/index.tsx
@@ -1,11 +1,7 @@
-import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/App';
 import { BrowserRouter } from 'react-router-dom';
 import TokenStore from './store/TokenStore';
-import CartStore from './store/CartStore';
-import UserStore from './store/UserStore';
-import CarStore from './store/CarStore';
 import { create } from 'mobx-persist';
 
 const hydrate = create({
