+++ src/pages/Catalog/CarCard/CarCard.tsx
@@ -1,0 +1,33 @@
+import CartStore from "store/CartStore";
+import classes from "./CarCard.module.css";
+import ReadonlyCar from "types/ReadonlyCar";
+
+
+function CarCard({ car }: { car: ReadonlyCar }) {
+
+    return (
+        <div className={classes.card}>
+            <div className={classes.items}>
+                <div>
+                    №{car.id}
+                </div>
+                <div>
+                    {car.brand} {car.model} {car.color}
+                </div>
+                <div>
+                    {car.price} рублей
+                </div>
+                <button className={classes.cartButton}
+                    onClick={() => CartStore.addToCart(car)}>
+                    В корзину
+                </button>
+            </div>
+            <img
+                className={classes.image}
+                src={`${process.env.PUBLIC_URL}/png/lada.png`}>
+            </img>
+        </div>
+    );
+}
+
+export default CarCard;
